Refactor to serve rooms based on room id
