--- client/components/Calendar.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { CalendarHeader } from './CalenderHeader.jsx';
-
-function Calendar(props) {
-    return (
-      <div className={props.calendar.length === 6 ? 'calContainer calXL' : 'calContainer'}>
-        <div >
-          <CalendarHeader buildCalendarRowForSelectedMonth={props.buildCalendarRowForSelectedMonth} determineStructureOfCalendarForSelectedMonth={props.determineStructureOfCalendarForSelectedMonth} onCalendarChangeClick={props.onCalendarChangeClick} selectDates={props.selectDates}/>
-        </div>
-        <div className='calInner'>
-          {props.calendar}
-        </div>
-      </div>
-  )
-}
-
-export {
-  Calendar
-}
-
-
-
-
