--- client/components/PriceRatingBanner.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function PriceRatingBanner(props) {
-  return (
-    <div>holder for price ratings banner</div>
-  )
-}
-
-export {
-  PriceRatingBanner
-}
