--- client/components/styles.css
@@ -1,214 +1,0 @@
-/* GENERAL */
-* {
-    margin: 0;
-    padding: 0;
-}
-html {
-    box-sizing: border-box;
-    font-family: Circular, -apple-system, BlinkMacSystemFont, Roboto, Helvetica Neue, sans-serif;
-}
-
-/* BOOKING CONTAINER */
-/* .BCstyle {
-    margin-left: 45px;
-    width: 376px;
-    z-index: 3;
-} */
-
-/* BOOKING */
-.bookingOuter {
-    position: fixed;
-    top: 75px;
-    /* temp reposition just to see both components during developement */
-    margin-left: 500px;
-    /* margin-left: 45px; */
-    width: 376px;
-    z-index: 3;
-}
-.bookingBorder {
-    padding-left: 24px;
-    padding-right: 24px;
-    margin: 0px;
-    border: 1px solid #e4e4e4;
-    background-color: #ffffff;
-    display: block;
-}
-.bookingSpacer {
-    margin-top: 16px;
-    margin-bottom: 24px;
-    display: block;
-}
-.bookingContentArea {
-    line-height: 1.43;
-    color: #484848;
-}
-.bookingForm {
-
-}
-.bookingFormStandardFields {
-
-}
-.bookingFormFieldsTopMargin {
-    margin-top: 16px;
-    margin-bottom: 8px;
-}
-/* DATE RANGE PICKER */
-.DRPcheckinCheckout {
-    width: 326px;
-    height: 42px;
-    font-weight: 600;
-}
-.DRPcheckinCheckoutBorder {
-    display: table;
-    table-layout: fixed;
-    border: 1px solid rgb(235, 235, 235);
-    border-radius: 2px;
-}
-.DRPcheckinCheckoutTR {
-    display: table-row;
-}
-.DRPcheckinTableCell {
-    display: table-cell;
-
-}
-.DRPcheckoutTableCell {
-    display: table-cell;
-}
-.DRPcheckinCheckoutPad {
-    font-weight: normal;
-    font-size: 17px;
-    line-height: 24px;
-    color: rgb(117, 117, 117);
-    position: relative;
-    display: inline-block;
-    width: 100%;
-    vertical-align: middle;
-    margin: 0px;
-    padding: 8px;
-    background: rgb(255, 255, 255)
-
-}
-.DRPcheckoutPad {
-
-}
-/* CALENDAR ROW */
-.calRow {
-    display: table;
-    box-sizing: border-box;
-    width: 100%;
-}
-/* CALENDAR CELL */
-.calDayBooked {
-    color: rgb(216, 216, 216) !important;
-    margin: 0px;
-    text-decoration: line-through;
-}
-.calDay {
-    display: inline-block;
-    text-align: center;
-    vertical-align: middle;
-    box-sizing: border-box; 
-    line-height: 12px;
-    font-size: 14px;
-    width: 41px; 
-    height: 40px;
-    padding: 13px 0;
-    color: #484848;
-    border: 1px double rgb(228, 231, 231);
-}
-.calDayNull {
-    border: none;
-}
-/* CALENDAR */
-.calInner {
-    text-align: center;
-    position: relative;
-    top: 86px;
-    left: 1px;
-    width: 100%;
-}
-.calContainer {
-    box-sizing: border-box;
-    width: 332px;
-    height: 308px;
-    border: 1px solid rgb(228, 228, 228);
-    display: inline-block;
-    border-radius: 3px;
-    position: absolute;
-}
-.calXL {
-    height: 348px !important;
-}
-/* CALENDAR HEADER */
-.calHead {
-    width: 288px;
-    height: 84px;
-    display: inline-block;
-    text-align: center;
-    font-size: 11.9px;
-    line-height: 1.43;
-    padding: auto 0;
-}
-.calHeadBtn {
-    border: 1px solid #757575;
-    border-radius: 3px;
-    position: absolute;
-    top: 18px;
-}
-.calHeadMonthYear {
-    font-size: 11.9px;
-    width: 41px; 
-    height: 20px;
-    line-height: 1.43;
-    display: inline-block;
-    text-align: center;
-    color: rgb(117, 117, 117);
-}
-.calHeadLeft {
-    left: 22px;
-}
-.calHeadRight {
-    right: 22px;
-}
-ul {
-    top: 62px;
-}
-li {
-    display: inline-block;
-}
-strong {
-    font-size: 18px;
-    display: inline-block;
-    vertical-align: middle;
-    color: rgb(72, 72, 72);
-}
-.calHeadBanner {
-    display: block;
-    text-align: center;
-    height: 25px;
-    width: 288px;
-    padding-top: 22px;
-    padding-bottom: 37px;
-}
-.calHeadWide {
-    position: absolute;
-    left: 21px;
-    text-align: center;
-
-}
-.calHeadDays {
-    box-sizing: border-box;
-    display: block;
-    position: absolute;
-    top: 64px;
-    left: 9.5px;
-    text-align: center;
-    padding: 0px 13px;
-    width: 313px;
-    height: 22px;
-}
-.svg {
-    padding: 6px 9px;
-    display: block;
-    color: rgb(117, 117, 117);
-}
