--- client/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './components/styles.css';
-import { BookingContainer } from './components/BookingContainer.jsx';
-
-window.BookingContainer = BookingContainer;
