+++ index.js
@@ -1,20 +1,21 @@
 const express = require('express');
 const parser = require('body-parser');
-// const path = require('path');
+const path = require('path');
 
 const port = 9000;
 const app = express();
 const { db } = require('./db/index');
 const { Booking } = require('./db/models/booking');
 
-app.use(express.static('public'));
+app.use('/rooms/:id/', express.static(path.join(__dirname, 'client/dist')));
 app.use(parser.urlencoded({ extended: true }));
 app.use(parser.json());
 
-app.post('/rooms/:room_id', (req, res) => {
-  let roomId = req.params.room_id;
+app.post('/api/rooms/:room_id', (req, res) => {
+  const roomId = req.params.room_id;
+  
   const genRandAvailScore = () => (Math.floor(Math.random() * 100));
-  let temp = {
+  const temp = {
     room_id: roomId,
     checkin: req.body.checkin,
     checkout: req.body.checkout,
@@ -30,7 +31,7 @@ app.post('/rooms/:room_id', (req, res) => {
   });
 });
 
-app.get('/rooms/:room_id', (req, res) => {
+app.get('/api/rooms/:room_id', (req, res) => {
   // eslint-disable-next-line prefer-const
   let roomId = req.params.room_id;
   Booking.findOne({ room_id: roomId }, 'availabilityScore', (err, id) => {
