--- public/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Airbnb bookings module</title>
-    <link rel="shortcut icon" href="#" />
-    <!-- <meta name="viewport" content="width=device-width, initial-scale=1"/> -->
-</head>
-<body>
-    <script crossorigin src="https://unpkg.com/react@16/umd/react.development.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
-    <script src="http://localhost:9000/app.js"></script>
-    <div id="BookingContainer"></div>
-    <script>
-      ReactDOM.render(
-        React.createElement(BookingContainer),
-        document.getElementById('BookingContainer')
-      );
-    </script>
-
-    
-</body>
-</html>
